requirements.txt
# Core MuSR dependencies
openai>=1.70.0
transformers>=4.51.0
tqdm>=4.66.0
numpy>=1.26.0
redis>=5.0.0
python-dotenv>=1.0.0

# CrewAI and LangChain for multi-agent reasoning
crewai>=0.201.0
langchain>=0.3.25
langchain-openai>=0.3.16
langchain-core>=0.3.58

# Optional dependencies for advanced features
# Uncomment if you need HuggingFace model support
# torch>=2.0.0
# accelerate>=0.20.0

# Development and testing (optional)
# pytest>=7.0.0
# black>=23.0.0